test(ai): extract isEarlyColonialExpansion group to sibling file (Refs #2509)
Move the newly added `isEarlyColonialExpansion` group out of
`observer_goal_phase_test.dart` into a focused sibling
`observer_goal_phase_is_early_colonial_expansion_test.dart` so the
parent suite stays under the 1000-non-comment-line cap
(`repo.dart_file_non_comment_line_size`, SPEC/program/repo-lint.md).

All seven cases (3 positive, 3 negative, 1 determinism) are preserved
unchanged; a one-line breadcrumb in the parent file points readers to
the new home. Refs #2509 — does not auto-close.
